+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG for ASAB WebUI Components
 
+## 27.1.1
+
+- Upgrade to react router v7, upgrade other "out of date" dependencies (#52)
+
+## 26.1.8
+
+- Implement new `DataTableAdvCustomFilter` with `CustomPill` feature for `DataTable2` (#38)
+
 ## 26.1.7
 
 - Extended the condition for DateTime in `validateDateTime`, whether the string is bigint. (#53)